.claude/skills/ai-content/video-automation/requirements.txt
# AI Video Automation Dependencies

# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# Optional: For advanced features
# openai>=1.12.0           # If using OpenAI Python SDK instead of raw requests
# pillow>=10.0.0           # Image manipulation
# moviepy>=1.0.3           # Alternative video processing (heavier than FFmpeg)
# anthropic>=0.18.0        # If using Claude for script generation
